Move call to registry to light module
